BUG: Lock problem caused invalid data
When releasing a lock for an item; if processing had taken too long or
if memcached dissappears and couldn't be connected to the message was
previously marked as needing redelivery, even though it may already have
been written properly to cassandra. This is bad news bears and almost
certainly has resulted in bad data on production since we had plenty of
problems with memcached with the memory leaks.

Reviewed-by: Paul Paradise <paul.paradise@socrata.com>