BUG: Memcache locking sucks.
And balboa cannot keep up with load events.

This REPLACES balboa memcache locking with a single-process-only lock
system that does not poll.  This will dramatically increase throughput
when a pile of events for the same row come in at once, at the cost of
requiring only one balboa process!

QA: No

Reviewed-by: aiden.scandella@socrata.com