This commit embodies a weekend when I (Mike Hotan) got frustrated with Balboa and attempted to fix too many problems at one time.
Background and Issues:
Balboa is a living example that Metrics system cannot be complete hacks like some software.  Balboa was not built to be retractable.  Retractable in the sense that any dependent systems that are directly using balboa are extremely coupled to its current implementation.  IE. When Metrics are wrong no one f'n thought about how to fix it.

Potential fixes and going forward:
Balboa components must be tested.  F showing your code is correct because it follows some theoretical models.  We must at the least test our business logic.  Software and Systems must be written with test and retractability in mind.  Can something be tested? If we find a problem later is there a framework in place to test the problem.

The result of my rathole and frustration:
* Make Date internal members in DataRange object immutable
* Standardize logging away from Apache Commons Logging to SLF4J + Log4j
* Reorganize poorly named and structured packages and namespace
* Rename KafkaCodec to BalboaCodec to identify specific dependency
* Standardize Logging into a central class in Common package.
* Added spurious documentation. JavaDoc
* Added a Period Comparator class to compare enum values.  I was wrong enums are already comparable: http://stackoverflow.com/questions/8007108/java-sorting-based-on-enum-constants
* Added Metrics Test Grab bag for testing to common package.
* Added MetricsServiceSpec and MetricsRestSpec.
* Cleaned up SBT Build settings.