Address cross compiling problems
balboa-client-* must be cross compiled for Scala 2.10 and 2.11 because
there are some projects of each version consuming them.

balboa-http can not be cross compiled for Scala 2.11 because the
socrata-http dependency is only in Scala 2.10.

The way our build system works it invokes all sbt tasks from the root
project:

    sbt +balboa-client-dispatcher/compile

When `sbt` is invoked this way, it uses the crossScalaVersions setting
from the root project, not from the project being compiled.

In the end we fixed this by munging the build job on Jenkins to pass an
additional parameter to sbt, thusly:

    sbt 'set crossScalaVersions = List("2.10.6", "2.11.7")' \
        +balboa-client-dispatcher/compile