NO-JIRA: Fix broken integration test script (#194)
Prior to this change, the balboa-http integration test script would fail
due to it being unable to determine the docker0 network interface IP
address. This appears to be due to changes in the output of ifconfig in
Ubuntu 18.04.

This change determines the IP address by using docker inspect with an
output formatter instead of parsing the output of ifconfig.