EN-36770: Periodically kill balboa-service-jms (#193)
Prior to this change, it is speculated that a long-running
balboa-service-jms instance could get into a state that would cause an
accumulation of messages in the Metrics2012 and Metrics2012.DLQ ActiveMQ
queues, eventually paging the on-call engineer and requiring manual
intervention to restart and scale the balboa-service-jms instances.

This change modifies the run script for balboa-service-jms' docker
container to automatically kill the main process, and thus the whole
container, approximately every 24 hours.  At this point Marathon will
replace the container with a fresh one, preventing any one container
from living longer than approximately 24 hours.